+++ test/support/mock_http.ex
@@ -20,7 +20,7 @@ defmodule OcapRpcTest.MockHttp do
 
     data =
       case method do
-        "eth_getBalance" -> TestUtils.user_balance() |> encode_ether()
+        "eth_getBalance" -> TestUtils.user_balance() |> Converter.to_hex()
         "eth_getCode" -> ""
         "eth_getBlockByHash" -> get_block(params)
         "eth_getBlockByNumber" -> get_block(params)
@@ -32,6 +32,8 @@ defmodule OcapRpcTest.MockHttp do
         "eth_gasPrice" -> TestUtils.gas_price()
         "eth_call" -> eth_call(params)
         "eth_getLogs" -> eth_log(params)
+        "trace_block" -> TestUtils.block_trace()
+        "trace_transaction" -> TestUtils.tx_trace()
         _ -> throw("error")
       end
 
@@ -50,13 +52,11 @@ defmodule OcapRpcTest.MockHttp do
   defp get_tx(_), do: TestUtils.tx()
 
   defp eth_call([%{data: _, from: _, to: @abt_addr}]),
-    do: TestUtils.user_balance() |> encode_ether()
+    do: TestUtils.user_balance() |> Converter.to_hex()
 
   defp eth_call([%{data: _, to: @abt_addr}]), do: TestUtils.abt_supply()
 
   defp eth_log([%{topics: _, address: @abt_addr}]), do: [TestUtils.tx()]
-
-  defp encode_ether(data), do: Converter.to_hex(data * Converter.ether())
 
   defp encode_resp(data, false) do
     %{"id" => 1, "result" => data}